refactor: move kyverno edges to kyverno file
ref: https://redhat.atlassian.net/browse/ACM-30858
