+++ pkg/transforms/kyverno_test.go
@@ -5,6 +5,7 @@ package transforms
 import (
 	"testing"
 
+	"github.com/stolostron/search-collector/pkg/config"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
@@ -395,6 +396,198 @@ func TestTransformKyvernoValidationAction(t *testing.T) {
 			AssertEqual("background", node.Properties["background"], tt.expectedBackground, t)
 			AssertEqual("admission", node.Properties["admission"], tt.expectedAdmission, t)
 			AssertEqual("severity", node.Properties["severity"], tt.expectedSeverity, t)
+		})
+	}
+}
+
+func TestKyvernoPolicyEdges(t *testing.T) {
+	// Establish the config
+	config.InitConfig()
+
+	configmap := &unstructured.Unstructured{Object: map[string]interface{}{
+		"apiVersion": "v1",
+		"kind":       "ConfigMap",
+		"metadata": map[string]interface{}{
+			"name":      "zk-kafka-address",
+			"uid":       "18b016fe-1931-4e80-95d1-d51d3b936e24",
+			"namespace": "test2",
+			"labels": map[string]interface{}{
+				"app.kubernetes.io/managed-by":          "kyverno",
+				"generate.kyverno.io/policy-name":       "zk-kafka-address",
+				"generate.kyverno.io/policy-namespace":  "",
+				"generate.kyverno.io/rule-name":         "k-kafka-address",
+				"generate.kyverno.io/trigger-group":     "",
+				"generate.kyverno.io/trigger-kind":      "Namespace",
+				"generate.kyverno.io/trigger-namespace": "",
+				"generate.kyverno.io/trigger-uid":       "12345",
+				"generate.kyverno.io/trigger-version":   "v1",
+				"somekey":                               "somevalue",
+			},
+		},
+	}}
+
+	configmapTwo := &unstructured.Unstructured{Object: map[string]interface{}{
+		"apiVersion": "v1",
+		"kind":       "ConfigMap",
+		"metadata": map[string]interface{}{
+			"name":      "hello-cofigmap",
+			"uid":       "777016fe-1931-4e80-95d1-d51d3b936e24",
+			"namespace": "test2",
+			"labels": map[string]interface{}{
+				"app.kubernetes.io/managed-by":          "kyverno",
+				"generate.kyverno.io/policy-name":       "kyverno-policy-test",
+				"generate.kyverno.io/policy-namespace":  "test2",
+				"generate.kyverno.io/rule-name":         "k-kafka-address",
+				"generate.kyverno.io/trigger-group":     "",
+				"generate.kyverno.io/trigger-kind":      "Namespace",
+				"generate.kyverno.io/trigger-namespace": "",
+			},
+		},
+	}}
+
+	configmapThree := &unstructured.Unstructured{Object: map[string]interface{}{
+		"apiVersion": "v1",
+		"kind":       "ConfigMap",
+		"metadata": map[string]interface{}{
+			"name":      "generated-secret",
+			"uid":       "888016fe-1931-4e80-95d1-d51d3b936e24",
+			"namespace": "test4",
+			"labels": map[string]interface{}{
+				"app.kubernetes.io/managed-by":          "kyverno",
+				"generate.kyverno.io/policy-name":       "generate-secret-policy",
+				"generate.kyverno.io/policy-namespace":  "",
+				"generate.kyverno.io/trigger-group":     "",
+				"generate.kyverno.io/trigger-kind":      "Namespace",
+				"generate.kyverno.io/trigger-namespace": "",
+			},
+		},
+	}}
+
+	configmapFour := &unstructured.Unstructured{Object: map[string]interface{}{
+		"apiVersion": "v1",
+		"kind":       "ConfigMap",
+		"metadata": map[string]interface{}{
+			"name":      "generated-config",
+			"uid":       "999016fe-1931-4e80-95d1-d51d3b936e24",
+			"namespace": "test3",
+			"labels": map[string]interface{}{
+				"app.kubernetes.io/managed-by":          "kyverno",
+				"generate.kyverno.io/policy-name":       "generate-configmap-policy",
+				"generate.kyverno.io/policy-namespace":  "test4",
+				"generate.kyverno.io/trigger-group":     "",
+				"generate.kyverno.io/trigger-kind":      "Pod",
+				"generate.kyverno.io/trigger-namespace": "test4",
+			},
+		},
+	}}
+
+	kyvernoPolicy := &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "kyverno.io/v1",
+			"kind":       "Policy",
+			"metadata": map[string]interface{}{
+				"name":      "kyverno-policy-test",
+				"namespace": "test2",
+				"uid":       "777738fa-0591-44da-8a06-98ea7d74a7f7",
+			},
+		},
+	}
+
+	kyvernoClusterpolicy := &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "kyverno.io/v1",
+			"kind":       "ClusterPolicy",
+			"metadata": map[string]interface{}{
+				"name": "zk-kafka-address",
+				"uid":  "8fc338fa-0591-44da-8a06-98ea7d74a7f7",
+			},
+		},
+	}
+
+	kyvernoGeneratingPolicy := &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "policies.kyverno.io/v1",
+			"kind":       "GeneratingPolicy",
+			"metadata": map[string]interface{}{
+				"name": "generate-secret-policy",
+				"uid":  "9fc338fa-0591-44da-8a06-98ea7d74a7f8",
+			},
+		},
+	}
+
+	kyvernoNamespacedGeneratingPolicy := &unstructured.Unstructured{
+		Object: map[string]interface{}{
+			"apiVersion": "policies.kyverno.io/v1",
+			"kind":       "NamespacedGeneratingPolicy",
+			"metadata": map[string]interface{}{
+				"name":      "generate-configmap-policy",
+				"namespace": "test4",
+				"uid":       "afc338fa-0591-44da-8a06-98ea7d74a7f9",
+			},
+		},
+	}
+
+	nodes := []Node{
+		KyvernoPolicyResourceBuilder(kyvernoPolicy).BuildNode(),
+		KyvernoPolicyResourceBuilder(kyvernoClusterpolicy).BuildNode(),
+		KyvernoPolicyResourceBuilder(kyvernoGeneratingPolicy).BuildNode(),
+		KyvernoPolicyResourceBuilder(kyvernoNamespacedGeneratingPolicy).BuildNode(),
+		GenericResourceBuilder(configmap).BuildNode(),
+		GenericResourceBuilder(configmapTwo).BuildNode(),
+		GenericResourceBuilder(configmapThree).BuildNode(),
+		GenericResourceBuilder(configmapFour).BuildNode(),
+	}
+	nodeStore := BuildFakeNodeStore(nodes)
+
+	tests := []struct {
+		name         string
+		resourceUID  string
+		expectedDest string
+		expectedKind string
+	}{
+		{
+			name:         "ClusterPolicy",
+			resourceUID:  "local-cluster/18b016fe-1931-4e80-95d1-d51d3b936e24",
+			expectedDest: "local-cluster/8fc338fa-0591-44da-8a06-98ea7d74a7f7",
+			expectedKind: "ClusterPolicy",
+		},
+		{
+			name:         "namespaced Policy",
+			resourceUID:  "local-cluster/777016fe-1931-4e80-95d1-d51d3b936e24",
+			expectedDest: "local-cluster/777738fa-0591-44da-8a06-98ea7d74a7f7",
+			expectedKind: "Policy",
+		},
+		{
+			name:         "GeneratingPolicy",
+			resourceUID:  "local-cluster/888016fe-1931-4e80-95d1-d51d3b936e24",
+			expectedDest: "local-cluster/9fc338fa-0591-44da-8a06-98ea7d74a7f8",
+			expectedKind: "GeneratingPolicy",
+		},
+		{
+			name:         "NamespacedGeneratingPolicy",
+			resourceUID:  "local-cluster/999016fe-1931-4e80-95d1-d51d3b936e24",
+			expectedDest: "local-cluster/afc338fa-0591-44da-8a06-98ea7d74a7f9",
+			expectedKind: "NamespacedGeneratingPolicy",
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Log("Test Kyverno " + tt.name)
+			edges := CommonEdges(tt.resourceUID, nodeStore)
+			if len(edges) != 1 {
+				t.Fatalf("Expected 1 edge but got %d", len(edges))
+			}
+
+			expectedEdge := Edge{
+				EdgeType:   "generatedBy",
+				SourceUID:  tt.resourceUID,
+				DestUID:    tt.expectedDest,
+				SourceKind: "ConfigMap",
+				DestKind:   tt.expectedKind,
+			}
+
+			AssertDeepEqual("edge", edges[0], expectedEdge, t)
 		})
 	}
 }
